requirements.txt
python-dotenv
fastapi
uvicorn
langchain
openai
ipykernel
langchain-openai
langchain-community
pypdf